Fix SunVersion selftest failed coredump
When building with the --debug config option, a selftest is run to
validate the Java version sanity checking code.  The version syntax
is:
  // 0 = major, 1 = minor, 2 = maintenance, 3 = update
with the first three parts delimted by '.' and the last part by '_'.

Originally, the update part only allowed two digits, but it was
later extended to three.
  //1.4.1_01-, 1.4.1_01a, the numerical part may only be 2 chars.
  //1.7.0_161, 1.8.0_162, it's as long as 3 chars in later Java version.

The code didn't actually accept three digit update values until this
commit:
  commit 3bd2d6aed629c4323ea9e8426acfb793eb9046fd
  Author: Damjan Jovanovic <damjan@apache.org>
  Date:   Sun Apr 15 15:00:46 2018 +0000

      Allow the Java version suffix (eg. the 162 in 1.8.0_162) to be
      3 digits long.

      Patch by: me

Unfortunately, this breaks the selftest, which has a list of valid
versions that should be accepted and another list that should be
rejected.  One of the bad values is 1.5.0_010.  After the above commit
this version was no longer rejected, which cause a failed assert and
core dump;
  Error: File .../main/jvmfwk/plugins/sunmajor/pluginlib/sunversion.cxx, Line 375: SunVersion selftest failed

Fix by adding a three digit update version to the valid versions list
and changing the bad list to have a four digit update.

Merge this change along with 3bd2d6aed629c4323ea9e8426acfb793eb9046f